added a conftest.py to pytables subdir
